Moved GoogleAnalytics to /analytics
