+++ src/analytics/index.js
@@ -1,0 +1,5 @@
+import GoogleAnalytics from './GoogleAnalytics'
+
+export {
+    GoogleAnalytics
+}
