+++ src/seo/index.js
@@ -1,7 +1,5 @@
 import MetaTags from './MetaTags';
-import GoogleAnalytics from './GoogleAnalytics'
 
 export {
     MetaTags,
-    GoogleAnalytics
 }
